+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Compatibility:
 Changes:
 
 * `TRUFFLERUBY_RESILIENT_GEM_HOME` has been removed. Unset `GEM_HOME` and `GEM_PATH` instead if you need to.
+* `Truffle::System.full_memory_barrier`, `Truffle::Primitive.logical_processors`, and  `Truffle::AtomicReference` have been removed.
 
 Performance: